# docs/swagger.yaml
basePath: /
definitions:
...
  common.BasicResponse:
    properties:
      message:
        type: string
    type: object
...
  common.GeneralResponse:
    properties:
      message:
        type: string
      result: {}
    type: object
  common.PaginatedResponse:
    properties:
      metadata:
        $ref: '#/definitions/pagination.Metadata'
      result:
        items: {}
        type: array
    type: object
  company.Company:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      region:
        type: string
      updatedAt:
        type: string
    type: object
...
  employee.Employee:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      phonenumber:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
...
  outlet.Outlet:
    properties:
      address:
        type: string
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  pagination.Metadata:
    properties:
      count:
        type: integer
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  user.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        enum:
        - superadmin
        - owner
        - employee
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  wage.Wage:
    properties:
      amount:
        type: integer
      code:
        type: string
      company:
        $ref: '#/definitions/company.Company'
      createdAt:
        type: string
      date:
        type: string
      employee:
        $ref: '#/definitions/employee.Employee'
      id:
        type: integer
      notes:
        type: string
      outlet:
        $ref: '#/definitions/outlet.Outlet'
      status:
        enum:
        - approved
        - accepted
        - canceled
        type: string
      type:
        enum:
        - debit
        - credit
        type: string
      updatedAt:
        type: string
    type: object
  wage.WageDTO:
    properties:
      amount:
        minimum: 0
        type: integer
      company:
        type: integer
      date:
        type: string
      employee:
        type: integer
      notes:
        type: string
      outlet:
        type: integer
      type:
        enum:
        - debit
        - credit
        type: string
    required:
    - amount
    - company
    - employee
    - outlet
    - type
    type: object
  wage.WageSummary:
    properties:
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      total:
        type: number
    type: object
info:
  contact:
    email: infamous0192@gmail.com
    name: Dwa Meizadewa
  description: Aplikasi Backend (API) untuk project Abude
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Abude Backend
  version: 2.0.0
paths:
...
  /api/wage:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: company
        type: integer
      - in: query
        name: employee
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: outlet
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/wage.Wage'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Wage
      tags:
      - Wages
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wage.WageDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/wage.Wage'
              type: object
      security:
      - JWT: []
      summary: Create Wage
      tags:
      - Wages
  /api/wage/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Wage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/wage.Wage'
              type: object
      security:
      - JWT: []
      summary: Delete Wage
      tags:
      - Wages
    get:
      consumes:
      - application/json
      parameters:
      - description: Wage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wage.Wage'
      security:
      - JWT: []
      summary: Get One Wage
      tags:
      - Wages
    put:
      consumes:
      - application/json
      parameters:
      - description: Wage ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wage.WageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/wage.Wage'
              type: object
      security:
      - JWT: []
      summary: Update Wage
      tags:
      - Wages
  /api/wage/{id}/cancel:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Wage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BasicResponse'
      security:
      - JWT: []
      summary: Cancel Wage
      tags:
      - Wages
  /api/wage/summary:
    get:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: query
        name: company
        type: integer
      - in: query
        name: employee
        type: integer
      - format: date-time
        in: query
        name: endDate
        type: string
      - description: Outlet ID
        in: query
        name: outlet
        type: integer
      - format: date-time
        in: query
        name: startDate
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - accepted
          - approved
          - canceled
          type: string
        name: status
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/wage.WageSummary'
            type: array
      security:
      - JWT: []
      summary: Get Wages Summary
      tags:
      - Wages
schemes:
- https
- http
securityDefinitions:
  JWT:
    description: Authorization for JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"